Updated Log In and Sign Up pages
I added variable checks to the log in and sign in pages. No longer will
people be allowed to try and sign in with no email, or no password, or
neither. I also added a check to make sure that the 2 passwords in the
sign up page match. Next step getting the errors from firebase auth,
and displaying them to the users