Added New coach information page
I added a new page that the new accounts will be taken to immediately
after auth. Asks for their first and last name swell as their school.
This is very basic page and will soon involve firebase database. Once
create account is clicked the users will be taken to the app homepage,
same as after users log in for the second+ time. Thank you and goodnight